fix bug when get message entity from database

+++ EmoChat/Controller/Extensions/DateExtension.swift
@@ -13,7 +13,7 @@ extension Date {
         return Int((self.timeIntervalSince1970 * 1000.0).rounded())
     }
     
-    init(milliseconds:Double) {
-        self = Date(timeIntervalSince1970: TimeInterval(milliseconds / 1000.0))
+    init(milliseconds:Int) {
+        self = Date(timeIntervalSince1970: TimeInterval(milliseconds / 1000))
     }
 }
